Added view/content
Added a UIStackView so that I can push on/push off the weight section of
the reps and such. It's nice because it auto-arranges the views for me.

On that note, I put all of the weight labels and sliders into a view so
that I could make them disappear/reappear more easily.

Added all of the different workouts that are in the P90X app, somewhat
needlessly but w/e.

Next up: try to make the weight labels/sliders fade in and out, use core
data to record history, or to do things in general, such as have all of
the information for all of the different workout sequences.