Added skipping functionality
All of the exercises with skipped indices are deleted both from the
database and from the last time you did stuff.

I will have to put in a check to make sure that the data that you're
getting will have all nine, and if not, it needs to add rows.